# optimize mixed str/int content