--- a/test/tests.py
+++ b/test/tests.py
@@ -298,7 +298,7 @@ class SATDataTests(BaseTest):
         """)
 
     def test_optimization_string_equality(self):
-        # just one key - skip block level optimization
+        # skip opimization - just one unique key
         self.assert_filter_has_n_results(4, """
             {
                 updated = "1714724220"
@@ -320,6 +320,23 @@ class SATDataTests(BaseTest):
                 updated = "1714724220"
                 dbn = "02M296"
                 not_existing_param ?= false
+            }
+        """)
+
+        # optimized, equality with mixed content, int + str
+        self.assert_filter_has_n_results(3, """
+            {
+                updated = 1714724220
+                zip_code = "10019"
+            }
+        """)
+
+        # skip optimization, integers only so "existence optimization"
+        # kicks in first
+        self.assert_filter_has_n_results(3, """
+            {
+                updated = 1714724220
+                zip_code = 10019
             }
         """)