--- a/test/tests.py
+++ b/test/tests.py
@@ -298,28 +298,11 @@ class SATDataTests(BaseTest):
         """)
 
     def test_optimization_string_equality(self):
-        # skip opimization - just one unique key
-        self.assert_filter_has_n_results(4, """
-            {
-                updated = "1714724220"
-                updated = "1714724220"
-            }
-        """)
-
         # optimized, equality alone
         self.assert_filter_has_n_results(3, """
             {
                 updated = "1714724220"
                 zip_code = "10019"
-            }
-        """)
-
-        # reproduce a bug - repeated key in AND expression
-        self.assert_filter_has_n_results(3, """
-            {
-                updated = "1714724220"
-                zip_code = "10019"
-                zip_code = "9999999999"
             }
         """)
 
@@ -340,29 +323,38 @@ class SATDataTests(BaseTest):
             }
         """)
 
-        # skip optimization, integers only so "existence optimization"
-        # kicks in first
-        self.assert_filter_has_n_results(3, """
-            {
-                updated = 1714724220
-                zip_code = 10019
-            }
-        """)
-
     def test_optimization_string_equality_integers_skipped(self):
+        # just one unique key
+        self.assert_filter_has_n_results(4, """
+            {
+                updated = "1714724220"
+                updated = "1714724220"
+            }
+        """)
+
+        # again just one unique key which is an optimization candidate
+        self.assert_filter_has_n_results(4, """
+            {
+                updated = "1714724220"
+                non_existing ?= false
+            }
+        """)
+
+        # repeated key in AND expression
+        self.assert_filter_has_n_results(0, """
+            {
+                updated = "1714724220"
+                zip_code = "10019"
+                zip_code = "9999999999"
+            }
+        """)
+
+        # integers only so "existence optimization" kicks in first
+        # ...and we never get to "equality optimization"
         self.assert_filter_has_n_results(3, """
             {
                 updated = 1714724220
                 zip_code = 10019
-            }
-        """)
-
-        # mixed content. one integer is enough to skip the optimization
-        self.assert_filter_has_n_results(1, """
-            {
-                updated = 1714724220
-                dbn = "02M296"
-                not_existing_param ?= false
             }
         """)