In addition to #126 this PR takes into account also integers and treats them same way as strings when it comes to optimization.
Normally, the "integers only" optimization would be applied as well, but it is preempted by the earlier [existence optimization](https://github.com/mediapredict/daffodil/blob/master/daffodil/hstore_predicate.pyx#L138), preventing it from being triggered.

Code excerpt for the record:
```
        if not breaks_existence_optimizer(children):
            return self.optimize_existence(children, sql_expr, "?&", "AND")

        if not breaks_equality_optimizer(children):
            return self.optimize_equality_and(children, sql_expr)
```
(inverted optimization ruin performance, should be avoided. Existence should be the first)
 
### Motivation
Vix: It turns out that this change speeds up the calculation by further 31% compared to the first (strings only) optimization step. 
All Non-Vix routers: from 321.8 seconds down to 239.2 seconds, which is 25% (relative to string only optimization) or 264.2 seconds down to 239.2 seconds which is only 9.5% faster.  

Conclusion this PR fixes the performance for non-Vix LFRs which got ruined by the first optimization. 

### Overall
Vix: 62% faster
Slowest vix: 27x faster
Non-vix: 9.5% faster