@igorkramaric 2024-05-04
Extra: 
Discovered a bug. In a "multiple string/integer equality AND expression" when using the same key twice, with different values the count of results should obviously be zero. However, optimized SQL may easily give result other than zero.

Optimization should not be skipped whenever there are any non-unique keys.